refactor(sync): make the merge rules testable and stop resetting favorite_modified_at
SyncMerger holds the merge rules that used to live on SyncService. They only
depend on the two SyncData documents plus SyncPreferences, so moving them out
removes the Android Context the tests had to mock. SyncService keeps the
transport-specific doSync and exposes the merger to its subclasses.

update_last_favorited_at_mangas gets a `WHEN new.favorite != old.favorite`
guard. mangasQueries.update always lists `favorite = coalesce(:favorite,
favorite)`, and SQLite's AFTER UPDATE OF favorite fires whenever the column
appears in the SET list, so every unrelated write reset favorite_modified_at
and it could no longer tell "the user favorited this" apart from "this row was
touched". Recreated in 41.sqm, which has not shipped yet.
